Add a simple JSON-NL input for a little more flexibility.
Allow input lines to be in a micro-JSON format like the following:

```
{"host":"https://139.162.1.157:443","hostname":"li843-157.members.linode.com"}
```

This will tell the runner to connnect to that IP address on that port,
but set the `Host` header to "li843-157.members.linode.com".

--

This also adds a new flag `--host-header` or `-hh` which will output the
custom host that was supplied to the target.